fixed AQICN models and finished AqicnApiManager implementation

+++ app/src/main/java/hpsaturn/pollutionreporter/api/AqicnH.java
@@ -5,5 +5,5 @@ package hpsaturn.pollutionreporter.api;
  */
 public class AqicnH {
 
-    public int v;
+    public double v;
 }
