+++ app/src/main/java/hpsaturn/pollutionreporter/api/AqicnTime.java
@@ -9,6 +9,6 @@ public class AqicnTime {
 
     public String tz;
 
-    public int v;
+    public double v;
 
 }
